django-backend/requirements.txt
adal==1.2.1
amqp==2.4.2
anyjson==0.3.3
asn1crypto==0.24.0
azure==4.0.0
azure-applicationinsights==0.1.0
azure-batch==4.1.3
azure-common==1.1.21
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.5
azure-datalake-store==0.0.45
azure-eventgrid==1.3.0
azure-graphrbac==0.40.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.0
azure-mgmt==4.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.50.0
azure-mgmt-batch==5.0.1
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-cdn==3.1.0
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==4.6.2
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-containerservice==4.4.0
azure-mgmt-cosmosdb==0.4.1
azure-mgmt-datafactory==0.6.0
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==1.0.0
azure-mgmt-devspaces==0.1.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==1.0.0
azure-mgmt-eventhub==2.6.0
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-iotcentral==0.1.0
azure-mgmt-iothub==0.5.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-logic==3.0.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementgroups==0.1.0
azure-mgmt-managementpartner==0.1.1
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.0
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-network==2.7.0
azure-mgmt-notificationhubs==2.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.1.0
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-rdbms==1.8.0
azure-mgmt-recoveryservices==0.3.0
azure-mgmt-recoveryservicesbackup==0.3.0
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.2.1
azure-mgmt-resource==2.2.0
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==2.1.0
azure-mgmt-servicebus==0.5.3
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-sql==0.9.1
azure-mgmt-storage==2.0.0
azure-mgmt-subscription==0.2.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.35.0
azure-nspkg==3.0.2
azure-servicebus==0.21.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.6
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
billiard==3.6.0.0
cachetools==3.1.1
celery==4.3.0
certifi==2018.1.18
cffi==1.12.3
chardet==3.0.4
click==6.7
colorama==0.4.1
cryptography==2.6.1
dataclasses==0.6
defusedxml==0.5.0
Django==2.0.3
django-braces==1.12.0
django-cors-headers==2.1.0
django-markdownx==2.0.28
django-modeltranslation==0.13.1
django-oauth-toolkit==1.0.0
django-rest-framework-social-oauth2==1.0.8
django-rest-hooks==1.5.0
djangorestframework==3.7.7
djongo==1.2.32
dnspython==1.16.0
eventlet==0.24.1
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
greenlet==0.4.15
httpie==1.0.2
httplib2==0.12.3
idna==2.6
isodate==0.6.0
kombu==4.5.0
Markdown==3.1.1
mock==3.0.5
monotonic==1.5
msrest==0.6.6
msrestazure==0.6.0
oauth2client==4.1.3
oauthlib==2.0.6
Pillow==6.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
Pygments==2.4.2
PyJWT==1.5.3
pymongo==3.8.0
PyMySQL==0.9.3
PyQt4==4.11.4
python-dateutil==2.8.0
python-dotenv==0.7.1
python3-openid==3.1.0
pytz==2017.3
redis==3.2.1
requests==2.18.4
requests-file==1.4.3
requests-oauthlib==0.8.0
rsa==4.0
selenium==3.141.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
sqlparse==0.2.4
tldextract==2.2.1
uritemplate==3.0.0
urllib3==1.22
vine==1.3.0
